Agregue las dependencias de bootstrap jquery y webjars-locator, Implemente las clases de bootstrap a las distintas paginas del proyecto, Corregí el AlumnoController y Alumno.html
