+++ TP5_Grupo11/src/main/java/ar/edu/unju/fi/controller/BecaController.java
@@ -1,4 +1,4 @@
-package ar.edu.unju.controller;
+package ar.edu.unju.fi.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -6,10 +6,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
 @RequestMapping("/inicio")
-public class CursoController {
+public class BecaController {
 
-	@GetMapping("/curso")
-	public String getCursoPage() {
-		return "Curso";
+	@GetMapping("/beca")
+	public String getBecaPage() {
+		return "Beca";
 	}
 }
