+++ TP5_Grupo11/src/main/java/ar/edu/unju/fi/controller/DocenteController.java
@@ -1,4 +1,4 @@
-package ar.edu.unju.controller;
+package ar.edu.unju.fi.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
