+++ TP5_Grupo11/src/main/java/ar/edu/unju/fi/model/Beca.java
@@ -1,4 +1,4 @@
-package ar.edu.unju.model;
+package ar.edu.unju.fi.model;
 
 public class Beca {
 	private int codigo;
