+++ TP5_Grupo11/src/main/java/ar/edu/unju/fi/model/Docente.java
@@ -1,30 +1,30 @@
-package ar.edu.unju.model;
+package ar.edu.unju.fi.model;
 
-public class Alumno {
-	private int dni;
+public class Docente {
+	private int legajo;
 	private String nombre;
 	private String apellido;
 	private String email;
 	private int telefono;
 	
-	
-	public Alumno(int dni, String nombre, String apellido, String email, int telefono) {
-		this.dni = dni;
+	public Docente(int legajo, String nombre, String apellido, String email, int telefono) {
+		this.legajo = legajo;
 		this.nombre = nombre;
 		this.apellido = apellido;
 		this.email = email;
 		this.telefono = telefono;
 	}
 
-	public Alumno() {
+	public Docente() {
+		super();
 	}
 
-	public int getDni() {
-		return dni;
+	public int getLegajo() {
+		return legajo;
 	}
 
-	public void setDni(int dni) {
-		this.dni = dni;
+	public void setLegajo(int legajo) {
+		this.legajo = legajo;
 	}
 
 	public String getNombre() {
@@ -58,7 +58,6 @@ public class Alumno {
 	public void setTelefono(int telefono) {
 		this.telefono = telefono;
 	}
-	
-	
+
 	
 }
