+++ TP5_Grupo11/src/main/resources/templates/layouts/footer.html
@@ -1,11 +1,15 @@
 <!DOCTYPE html>
 <html xmlns:th="http://www.thymeleaf.org">
-
+	<head>
+		<meta charset="UTF-8">
+	</head>
        <body>
-          <footer th:fragment="footer">
+          <footer th:fragment="footer" class="card-footer text-muted">
           
-             <p>2022. Portal de Capacitación – Av. Gral. Manuel Belgrano 789 – Teléfono (388) 999999 S. S. de Jujuy República Argentina</p>
-          
+             <p>2022. Portal de CapacitaciÃ³n â€“ Av. Gral. Manuel Belgrano 789 â€“ TelÃ©fono (388) 999999 S. S. de Jujuy RepÃºblica Argentina</p>
+          	 <script src="/webjars/jquery/jquery.min.js" type="text/javascript"></script>
+			 <script src="/webjars/popper.js/umd/popper.min.js" type="text/javascript"></script>
+			 <script src="/webjars/bootstrap/js/bootstrap.min.js" type="text/javascript"></script>
           
           </footer>>
